+++ Sources/ParseSwift/Objects/ParseInstallation+async.swift
@@ -16,7 +16,7 @@ public extension ParseInstallation {
      Fetches the `ParseInstallation` *aynchronously* with the current data from the server
      and sets an error if one occurs.
      - parameter includeKeys: The name(s) of the key(s) to include that are
-     `ParseObject`s. Use `["*"]` to include all keys. This is similar to `include` and
+     `ParseObject`s. Use `["*"]` to include all keys one level deep. This is similar to `include` and
      `includeAll` for `Query`.
      - parameter options: A set of header options sent to the server. Defaults to an empty set.
      - returns: Returns saved `ParseInstallation`.
@@ -128,7 +128,7 @@ public extension Sequence where Element: ParseInstallation {
      Fetches a collection of installations *aynchronously* with the current data from the server and sets
      an error if one occurs.
      - parameter includeKeys: The name(s) of the key(s) to include that are
-     `ParseObject`s. Use `["*"]` to include all keys. This is similar to `include` and
+     `ParseObject`s. Use `["*"]` to include all keys one level deep. This is similar to `include` and
      `includeAll` for `Query`.
      - parameter options: A set of header options sent to the server. Defaults to an empty set.
      - returns: Returns an array of Result enums with the object if a save was successful or a