+++ Package@5.1.swift
@@ -4,7 +4,11 @@ import PackageDescription
 
 let package = Package(
     name: "ParseSwift",
-    platforms: [.iOS(.v13), .macOS(.v10_15), .tvOS(.v13), .watchOS(.v6)],
+    platforms: [.iOS(.v13),
+                .macCatalyst(.v13),
+                .macOS(.v10_15),
+                .tvOS(.v13),
+                .watchOS(.v6)],
     products: [
         .library(
             name: "ParseSwift",