+++ Sources/ParseSwift/Objects/ParseInstallation.swift
@@ -508,7 +508,7 @@ extension ParseInstallation {
 
         var params: [String: String]?
         if let includeParams = include {
-            params = ["include": "\(includeParams)"]
+            params = ["include": "\(Set(includeParams))"]
         }
 
         return API.Command(method: .GET,