+++ Sources/ParseSwift/Objects/ParseUser.swift
@@ -916,7 +916,7 @@ extension ParseUser {
 
         var params: [String: String]?
         if let includeParams = include {
-            params = ["include": "\(includeParams)"]
+            params = ["include": "\(Set(includeParams))"]
         }
 
         return API.Command(method: .GET,